backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "buelldocs-backend"
version = "0.1.0"
description = "BuellDocs Document Generation Backend"
authors = [
    {name = "BuellDocs Team", email = "team@buelldocs.com"},
]
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "sqlalchemy==2.0.23",
    "asyncpg==0.29.0",
    "alembic==1.13.1",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.6",
    "passlib[bcrypt]==1.7.4",
    "weasyprint==61.2",
    "jinja2==3.1.2",
    "python-dateutil==2.8.2",
    "email-validator==2.1.0.post1",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "black==23.11.0",
    "isort==5.12.0",
    "flake8==6.1.0",
]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88